+++ src/main/java/com/challenge/vendingmachine/validator/FiveMultipleValidator.java
@@ -4,10 +4,10 @@ package com.challenge.vendingmachine.validator;
 import javax.validation.ConstraintValidator;
 import javax.validation.ConstraintValidatorContext;
 
-public class FiveMultipleValidator implements ConstraintValidator<FiveMultiple, Double> {
+public class FiveMultipleValidator implements ConstraintValidator<FiveMultiple, Long> {
 
     @Override
-    public boolean isValid(Double aDouble, ConstraintValidatorContext constraintValidatorContext) {
+    public boolean isValid(Long aDouble, ConstraintValidatorContext constraintValidatorContext) {
         return aDouble != null && aDouble % 5 == 0;
     }
 }
